refactor into proper pattern
